- Updated multiboot & emulator
git-svn-id: svn://localhost/mosa@1448 45b04e34-05f1-4d87-a130-40a8491d5c8b

+++ Mosa/Emulator/HelloWorld/PageFrameAllocator.cs
@@ -52,7 +52,7 @@ namespace Mosa.Kernel.Memory.X86
 				return;
 
 			for (uint index = 0; index < Multiboot.MemoryMapCount; index++) {
-				uint value = (uint)Multiboot.GetMemoryMapType(index);
+				byte value = Multiboot.GetMemoryMapType(index);
 
 				Mosa.HelloWorld.Screen.SetCursor(22, index);
 				Mosa.HelloWorld.Screen.Write(value);
